move pdf-folder-index to pages folder
