+++ src/app/app-routing.module.ts
@@ -7,7 +7,7 @@ import { ProjectsComponent } from './components/projects/projects.component';
 import { QuotesComponent } from './components/quotes/quotes.component';
 import { MainPageComponent } from './pages/main-page/main-page.component';
 import { PageNotFoundComponent } from "./components/page-not-found/page-not-found.component";
-import { PdfFolderIndexComponent } from "./components/pdf-folder-index/pdf-folder-index.component";
+import { PdfFolderIndexComponent } from "./pages/pdf-folder-index/pdf-folder-index.component";
 
 const routes: Routes = [
   {
