+++ src/app/pages/pdf-folder-index/pdf-folder-index.component.scss
